Separando arquivos de mat.h
