+++ include/math.hpp
@@ -2,8 +2,9 @@
 #define S3D_MATH_HPP
 
 #include "cppfn.h"
-#include "mat.h"
 #include "vec.h"
+
+#include "matrix/type.h"
 
 cppfn void s3d_translate(s3d_fmat4 mat, const s3d_fvec3 val);
 cppfn void s3d_scale(s3d_fmat4 mat, const s3d_fvec3 val);
