+++ include/matrix/copy.h
@@ -1,0 +1,8 @@
+#ifndef S3D_MATRIX_COPY_H
+#define S3D_MATRIX_COPY_H
+
+#include "matrix/type.h"
+
+void s3d_mat_copy(s3d_fmat4 dest, const s3d_fmat4 src);
+
+#endif
