+++ include/matrix/identity.h
@@ -1,0 +1,8 @@
+#ifndef S3D_MATRIX_IDENTITY_H
+#define S3D_MATRIX_IDENTITY_H
+
+#include "matrix/type.h"
+
+void s3d_identity(s3d_fmat4 mat);
+
+#endif
