+++ include/matrix/type.h
@@ -1,0 +1,6 @@
+#ifndef S3D_MATRIX_TYPE_H
+#define S3D_MATRIX_TYPE_H
+
+typedef float s3d_fmat4[4][4];
+
+#endif
