+++ src/matrix/identity.c
@@ -1,8 +1,7 @@
-#include "mat.h"
+#include "matrix/identity.h"
 
-#include<string.h>
-
-#include "vec.h"
+#include "matrix/type.h"
+#include "matrix/copy.h"
 
 void s3d_identity(s3d_fmat4 mat) {
     static const s3d_fmat4 id = {
@@ -14,7 +13,4 @@ void s3d_identity(s3d_fmat4 mat) {
 
     s3d_mat_copy(mat, id);
 }
- 
-void s3d_mat_copy(s3d_fmat4 dest, const s3d_fmat4 src) {
-    memcpy(dest, src, sizeof(s3d_fmat4));
-}
+
